requirements.txt
mysql-connector==2.1.6
requests==2.19.1
colorama==0.3.9
